ADL: Rename adl_v1.cpp/h to hires1.cpp/h
